requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg             # High-speed Postgres/Hologres driver
jinja2
python-multipart    # For login forms
pydantic-settings
passlib[bcrypt]     # Security
bcrypt<4.2
python-jose[cryptography]         # Token management
aiofiles            # Async file handling
python-dotenv
email-validator
